Implement pretty tone display

+++ src/pinyin/syllable.cpp
@@ -4,6 +4,6 @@
 
 std::ostream &operator<<(std::ostream &out, const Syllable &syl) {
 	return out << CONSONANT_STR[syl.consonant].str << VOWEL_STR[syl.vowel].str
-			   << syl.tone;
+			   << (int)syl.tone;
 }
